got rid of some clutter
