+++ seeker/src/Pages/Login.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+class Login extends React.Component{
+    constructor(){
+        super();
+        this.state={};
+    }
+
+    render(){
+        return(
+            <p>Hello world</p>
+        )
+    }
+}
+
+export default Login;
